requirements.txt
amqp==5.3.1
anyio==4.12.0
asgi-cors==1.0.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.4
boto3==1.42.13
botocore==1.42.13
celery==5.6.0
certifi==2025.11.12
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.12.0
Django==5.2.8
django-cors-headers==4.9.0
django-pytest==0.2.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
exceptiongroup==1.3.1
factory_boy==3.3.3
Faker==38.2.0
flower==2.0.1
h11==0.16.0
httptools==0.7.1
humanize==4.14.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg-binary==3.2.12
psycopg2==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
s3transfer==0.16.0
sentry-sdk==2.48.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
